Move pages into their own folder.
