+++ src/bulk_upload.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import BulkUploadPage from './components/BulkUploadPage';
-import ErrorBoundary from './components/ErrorBoundary';
+import ErrorBoundary from 'components/ErrorBoundary';
+import BulkUploadPage from 'pages/BulkUploadPage';
 
 const wrapper = document.getElementById('react-root');
 const element = (
