+++ src/cube_deck.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import CubeDeckPage from 'components/CubeDeckPage';
+import CubeDeckPage from 'pages/CubeDeckPage';
 
 const element = <CubeDeckPage {...reactProps} />;
 const wrapper = document.getElementById('react-root');
