+++ src/cube_deckbuilder.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import CubeDeckbuilderPage from 'components/CubeDeckbuilderPage';
+import CubeDeckbuilderPage from 'pages/CubeDeckbuilderPage';
 
 const wrapper = document.getElementById('react-root');
 wrapper ? ReactDOM.render(<CubeDeckbuilderPage {...reactProps} />, wrapper) : false;
