+++ src/cube_playtest.js
@@ -1,8 +1,8 @@
-import React, { Component, Fragment } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 
-import CubePlaytestPage from './components/CubePlaytestPage';
-import ErrorBoundary from './components/ErrorBoundary';
+import ErrorBoundary from 'components/ErrorBoundary';
+import CubePlaytestPage from 'pages/CubePlaytestPage';
 
 const wrapper = document.getElementById('react-root');
 const element = (
