+++ src/pages/CubeDeckPage.js
@@ -23,7 +23,6 @@ import { sortDeck } from 'util/Util';
 
 import CustomImageToggler from 'components/CustomImageToggler';
 import { DisplayContextProvider } from 'components/DisplayContext';
-import { subtitle } from 'components/CubeDraftPage';
 import DynamicFlash from 'components/DynamicFlash';
 import FoilCardImage from 'components/FoilCardImage';
 import { getCardColorClass } from 'components/TagContext';
@@ -31,6 +30,7 @@ import withAutocard from 'components/WithAutocard';
 import CommentEntry from 'components/CommentEntry';
 import CommentsSection from 'components/CommentsSection';
 import CubeLayout from 'layouts/CubeLayout';
+import { subtitle } from 'pages/CubeDraftPage';
 
 const AutocardItem = withAutocard(ListGroupItem);
